# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TaskEarn Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/transactions:
    get:
      operationId: getTransactions
      tags: [wallet]
      summary: Get transaction history
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /wallet/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [wallet]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "200":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Insufficient balance

  # ─── REFERRALS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, type, amount, status, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        status:
          type: string
        description:
          type: string
        method:
          type: ["string", "null"]
        createdAt:
          type: string
    WithdrawalInput:
      type: object
      required: [amount, method, accountDetails]
      properties:
        amount:
          type: number
        method:
          type: string
        accountDetails:
          type: string

    # ── Referrals ─────────────────────────────────────────────────────────────
...